refactor: walk through scroll hook (#113)
